+++ Changelog.md
@@ -1,3 +1,4 @@
 --------------------------Changelog for v1.9.4-beta1 of SecurityCraft--------------------------
 
 - Change: Items in SecurityCraft's creative tabs have been reordered to match vanilla's order and in general make more sense
+- Change: Sounds for nether wood reinforced blocks have been adjusted to match the vanilla variants